Partial update for beta 0.1.052
Things aren't currently working at the moment because the compiler won't
let me modify functions marked `#compiler`, which would be necessary to
get rid of all uses of inline assembly so that we don't get compiler
errors about that.

Also, I forgot an added file in the modules patch. Oops!